+++ lib/msal-react/src/packageMetadata.ts
@@ -1,3 +1,3 @@
 /* eslint-disable header/header */
 export const name = "@azure/msal-react";
-export const version = "1.4.2";
+export const version = "1.4.3";